+++ cached_path/schemes/s3.py
@@ -8,8 +8,8 @@ from typing import Optional, Tuple
 import boto3
 import botocore
 
-from cached_path.common import _split_cloud_path
-from cached_path.schemes.scheme_client import SchemeClient
+from ..common import _split_cloud_path
+from .scheme_client import SchemeClient
 
 
 class S3Client(SchemeClient):