+++ tests/cached_path_test.py
@@ -384,3 +384,26 @@ class TestCachedPathHf(BaseTestClass):
         )  # should resolve to option 2.
         assert path.is_dir()
         assert self.TEST_DIR in path.parents
+
+
+def beaker_available() -> bool:
+    try:
+        from beaker import Beaker, ConfigurationError  # type: ignore
+
+        try:
+            Beaker.from_env()
+            return True
+        except ConfigurationError:
+            return False
+    except (ImportError, ModuleNotFoundError):
+        return False
+
+
+class TestCachedPathBeaker(BaseTestClass):
+    @flaky
+    @pytest.mark.skipif(not beaker_available(), reason="Beaker not configured")
+    def test_cache_object(self):
+        path = cached_path("beaker://petew/cached-path-readme/README.md")
+        assert path.is_file()
+        meta = Meta.from_path(_meta_file_path(path))
+        assert meta.etag is not None