+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added optional support for `beaker://` URLs.
+
 ## [v1.2.0](https://github.com/allenai/cached_path/releases/tag/v1.2.0) - 2023-01-12
 
 ### Changed